validation update product and product image

+++ app/controllers/Product.js
@@ -28,12 +28,22 @@ exports.get_all_products = (req, res, next) => {
 }
 
 exports.create_data_products = (req, res, next) => {
-    Product.create({
-        _id: new mongoose.Types.ObjectId(),
-        name: req.body.name,
-        price: req.body.price,
-        image: req.file.path
-    }, (err, product) => {
+    if(req.file){
+        var data = {
+            _id: new mongoose.Types.ObjectId(),
+            name: req.body.name,
+            price: req.body.price,
+            image: `uploads/${req.file.filename}`
+        }
+    }else{
+        var data = {
+            _id: new mongoose.Types.ObjectId(),
+            name: req.body.name,
+            price: req.body.price
+        }
+    }
+    
+    Product.create(data, (err, product) => {
         if(err)
             return next(new Error(err))
 
@@ -42,6 +52,7 @@ exports.create_data_products = (req, res, next) => {
                 _id: product._id,
                 name: product.name,
                 price: product.price,
+                image: product.image,
                 requests_detail: {
                     type: "GET",
                     url: `http://localhost:3000/api/products/${product._id}`
@@ -80,26 +91,81 @@ exports.update_data_products = (req, res, next) => {
         else if (!products)
             return next(new Error("Product not found"))
         
-        fs.unlink(`./${products.image}`, (err) => {
-            if(err)
-                return next(new Error("Image not found"))
-
+        // check request file 
+        if(req.file){
+            if(products.image.split("uploads/")[1] !== "default_image.png"){
+                // delete file upload and update file upload
+                fs.unlink(`./${products.image}`, (err) => {
+                    if(err)
+                        return next(new Error("Image not found"))
+    
+                    Product.updateOne({_id: products._id}, {
+                        $set: {
+                            name: req.body.name,
+                            price: req.body.price,
+                            image: `uploads/${req.file.filename}`
+                        }
+                    }, (err) => {
+                        if(err)
+                            return next(new Error("Problem updated products"))
+                    
+                        res.status(200).json({
+                            products: {
+                                _id: products._id,
+                                name: req.body.name,
+                                price: req.body.price,
+                                image: `uploads/${req.file.filename}`,
+                                requests_detail: {
+                                    type: "GET",
+                                    url: `http://localhost:3000/api/products/${products._id}`
+                                }
+                            }
+                        })
+                    })
+                })
+    
+            }else{
+                Product.updateOne({_id: products._id}, {
+                    $set: {
+                        name: req.body.name,
+                        price: req.body.price,
+                        image: `uploads/${req.file.filename}`
+                    }
+                }, (err) => {
+                    if(err)
+                        return next(new Error("Problem updated products"))
+                
+                    res.status(200).json({
+                        products: {
+                            _id: products._id,
+                            name: req.body.name,
+                            price: req.body.price,
+                            image: `uploads/${req.file.filename}`,
+                            requests_detail: {
+                                type: "GET",
+                                url: `http://localhost:3000/api/products/${products._id}`
+                            }
+                        }
+                    })
+                })
+                
+            }
+        }else{
             Product.updateOne({_id: products._id}, {
                 $set: {
                     name: req.body.name,
                     price: req.body.price,
-                    image: `uploads/${req.file.filename}`
                 }
             }, (err) => {
                 if(err)
                     return next(new Error("Problem updated products"))
-               
+            
                 res.status(200).json({
                     products: {
                         _id: products._id,
                         name: req.body.name,
                         price: req.body.price,
-                        image: `uploads/${req.file.filename}`,
+                        image: products.image,
                         requests_detail: {
                             type: "GET",
                             url: `http://localhost:3000/api/products/${products._id}`
@@ -107,7 +173,8 @@ exports.update_data_products = (req, res, next) => {
                     }
                 })
             })
-        })
+
+        }
     })
 }
 
@@ -118,10 +185,27 @@ exports.delete_data_products = (req, res, next) => {
         else if(!products)
             return next(new Error("Product not found"))
 
-        fs.unlink(`./${products.image}`, (err) => {
-            if(err)
-                return next("Image not found")
+        // check file name in database
+        if(products.image.split("uploads/")[1] !== "default_image.png"){
+            // delete file upload and delete product
+            fs.unlink(`./${products.image}`, (err) => {
+                if(err)
+                    return next("Image not found")
 
+                Product.deleteOne({_id: products._id}, (err) => {
+                    if(err)
+                        return next(new Error("Problem deleted product"))
+        
+                    res.status(200).json({
+                        message: "Deleted products!",
+                        requests_all: {
+                            type: "GET",
+                            url: "http://localhost:3000/api/products"
+                        }
+                    })
+                })
+            })
+        }else{
             Product.deleteOne({_id: products._id}, (err) => {
                 if(err)
                     return next(new Error("Problem deleted product"))
@@ -134,6 +218,6 @@ exports.delete_data_products = (req, res, next) => {
                     }
                 })
             })
-        })
+        }
     })
 }
